backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9